Finishing wave animation
