+++ src/components/LoadingScreen/LoadingScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Image } from 'react-native';
+import { View, Image, Modal } from 'react-native';
 import { useTheme } from '@hooks';
 import { Text } from '@components';
 import styles from './LoadingScreen.styles';
@@ -8,17 +8,19 @@ import { LoadingScreenProps } from './LoadingScreen.types';
 const LoadingScreen: React.FC<LoadingScreenProps> = ({ title }) => {
 	const { colors } = useTheme();
 	return (
-		<View style={[styles.container, { backgroundColor: colors.background1 }]}>
-			<Image style={styles.image} source={require('../../../assets/wave.gif')} />
-			<View>
-				<Text type="h2" weight="bold" marginBottom={8} center>
-					{title}
-				</Text>
-				<Text type="p2" center>
-					This can take a few seconds...
-				</Text>
+		<Modal visible>
+			<View style={[styles.container, { backgroundColor: colors.background1 }]}>
+				<Image style={styles.image} source={require('../../../assets/wave.gif')} />
+				<View>
+					<Text type="h2" weight="bold" marginBottom={8} center>
+						{title}
+					</Text>
+					<Text type="p2" center>
+						This can take a few seconds...
+					</Text>
+				</View>
 			</View>
-		</View>
+		</Modal>
 	);
 };
